// apps/main/src/models/v2/iam/userWithRoleType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */

/**
 * The type of user - human user or service user
 */
export type UserWithRoleType = (typeof UserWithRoleType)[keyof typeof UserWithRoleType];

export const UserWithRoleType = {
  user: 'user',
  'service-user': 'service-user',
} as const;